MIE Phase A+B+C: doc update, test split, src split (83/83 pass)
Phase A — Documentation
- mie-architecture.md: rewrite §4.2 (MIED v2 format), §5 (5-mode table +
  tone-aware ranking), §8 (roadmap all Phase-1 items marked done, 83 tests);
  add §10 cross-platform C API design note.
- software-requirements.md: update §5 mode table to 5 modes, test count 83.

Phase B — Test split (1554-line monolith → 3 files + shared header)
- tests/test_helpers.h: shared TEntry, kev, kMODE/kSPACE/kOK/kBACK,
  build_single, build_two_zh, build_multi.
- tests/test_ime_basic.cpp: SmartMode, DirectMode, SymbolKeys,
  ContextLang, CandidateNav (36 tests).
- tests/test_ime_modes.cpp: ModeSwitch, ModeSeparation (12 tests).
- tests/test_ime_advanced.cpp: GreedyPrefix, AbbreviatedInput, SpaceTone,
  CompoundDisplay, SingleCharPriority, SmartEn, BackspaceTone1,
  DirectBopomofo, ToneSort (21 tests).
- tests/test_ime_logic.cpp: deleted.

Phase C — Source split (991-line monolith → 7 files + internal header)
- src/ime_internal.h: DirectEntry/kDirectTable/find_direct_entry,
  utf8_char_count, extract_tone_intent, tone_tier, is_direct_mode.
- src/ime_keys.cpp: kPhonemeTable, kDirect/Sym tables, all key_to_*/
  direct_*/sym_* methods.
- src/ime_search.cpp: run_search, build_merged, rebuild_input_buf,
  matched_prefix_display_bytes.
- src/ime_display.cpp: append/backspace/set_display, mode_indicator,
  compound_input_str, compound_input_bytes, matched_prefix_compound_bytes.
- src/ime_commit.cpp: do_commit, do_commit_partial, did_commit.
- src/ime_smart.cpp: process_smart.
- src/ime_direct.cpp: commit_sym_pending, process_sym_key, process_direct.
- src/ime_logic.cpp: trimmed to constructor, set_commit_callback,
  clear_input, process_key dispatcher (~110 lines).

All 83 C++ unit tests pass. No behaviour change.
