+++ firmware/mie/tests/CMakeLists.txt
@@ -13,7 +13,9 @@ FetchContent_MakeAvailable(googletest)
 add_executable(mie_tests
     test_trie_stub.cpp
     test_trie_searcher.cpp
-    test_ime_logic.cpp
+    test_ime_basic.cpp
+    test_ime_modes.cpp
+    test_ime_advanced.cpp
 )
 
 target_link_libraries(mie_tests
